requirements.txt
fastapi
uvicorn
python-multipart
starlette
timm
Pillow
gunicorn
torch
torchvision
numpy
scipy
